+++ basics/processing-instructions/steel/program/src/lib.rs
@@ -9,15 +9,11 @@ pub fn process_instruction(
 ) -> ProgramResult {
     msg!("Processing instruction");
 
-    let (_tag, instruction_data) = instruction_data
-        .split_first()
-        .ok_or(ProgramError::InvalidInstructionData)?;
+    let instruction = GoToTheParkData::try_from_bytes(instruction_data)?;
 
-    let instruction = GoToThePark::try_from_bytes(instruction_data)?;
+    msg!("Welcome to the park, {}", instruction.name());
 
-    msg!("Welcome to the park, {}", instruction.data.name());
-
-    if instruction.data.height() > 5 {
+    if instruction.height() > 5 {
         msg!("You are tall enough to ride this ride. Congratulations.");
     } else {
         msg!("You are NOT tall enough to ride this ride. Sorry mate.");
